refactor: standardize file naming
