+++ frontend/src/app.jsx
@@ -4,7 +4,7 @@ import "./scss/app.scss";
 import {Header} from "./layout/header";
 import {Lawrgminer} from "./components/miner/lawrgminer";
 import {Footer} from "./layout/footer";
-import {ReportSection} from "./components/report/reportSection";
+import {ReportSection} from "./components/report/report-section";
 
 export class App extends React.Component {
     render() {
