+++ frontend/src/components/miner/drag-and-drop.jsx
