+++ frontend/src/components/miner/file-upload.jsx
