+++ frontend/src/components/miner/mock-data.js
