+++ frontend/src/components/report/report-section.jsx
