+++ frontend/src/index.js
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 
 import "./index.scss";
-import {App} from "./App";
+import {App} from "./app";
 import reportWebVitals from "./reportWebVitals";
 
 ReactDOM.render(
